% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_into_html_and_pdf.r
\name{compile_rmd_multiple_outputs}
\alias{compile_rmd_multiple_outputs}
\title{Compile an R Markdown file to multiple outputs}
\usage{
compile_rmd_multiple_outputs(input_file, max_wait = 10)
}
\arguments{
\item{input_file}{The path to the R Markdown file to be compiled.}

\item{max_wait}{Max. waittime (minutes) until the pdf finishes printing.
Default = 10}
}
\value{
Invisibly returns NULL. The side effect of this function is that
it writes the compiled documents to disk (same folder as source).
}
\description{
This function takes an R Markdown file as input and generates three outputs:
an HTML document, a PDF-like HTML document using the pagedown package, and
a PDF document. If generating the PDF document fails (e.g., if Chrome is not
installed), the function will still generate the PDF-like HTML document.
}
\examples{
\dontrun{
compile_rmd_multiple_outputs("path/to/your_file.Rmd")
}

}
