% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.r
\docType{data}
\name{data}
\alias{data}
\title{Modified Iris Dataset}
\format{
A data frame with 150 rows and 9 variables:
\describe{
\item{species_no}{Numeric representation of species. It is a numeric vector of length 150 with no missing values and 3 distinct values.}
\item{petal_width}{These are the width of the petals. It is a numeric vector of length 150 with no missing values and 22 distinct values.}
\item{petal_length}{These are the length of the petals. It is a numeric vector of length 150 with no missing values and 43 distinct values.}
\item{sepal_width}{These are the width of the sepals. It is a numeric vector of length 150 with no missing values and 23 distinct values.}
\item{sepal_length}{These are the length of the sepals. It is a numeric vector of length 150 with no missing values and 35 distinct values.}
\item{species_char}{Character representation of the species. It is a factor with 150 levels and 3 distinct values.}
\item{date}{This is a date column to illustrate transformations. It is a Date vector of length 150 with no missing values and 150 distinct values.}
\item{newvar}{This is a new example variable, adding up the lengths. It is a numeric vector of length 150 with no missing values and 42 distinct values.}
\item{mock_ID}{It is a numeric vector of length 150 with no missing values and 21 distinct values.}
}
}
\usage{
data
}
\description{
This is the famous Iris dataset, often used for teaching and demonstrating
machine learning algorithms. The dataset consists of measurements of
150 iris flowers from three different species.
}
\details{
I've modified it to demonstrate basic dplyr style wrangling
}
\keyword{datasets}
