% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_location.r
\name{get_current_file_location}
\alias{get_current_file_location}
\title{Get Current File Location}
\usage{
get_current_file_location()
}
\value{
A character string with the directory path of the current file.
If no file path is found, a message is printed and NULL is returned.
}
\description{
This function returns the directory path of the current file in execution.
The function first checks if it is ran by another process
(if there's a command argument '--file' and retrieves its value).
If not, the function then attempts to get the path
from the active document in the RStudio source editor.
If this is also unsuccessful (for instance, if the function is being run
from a new, unsaved R script),
the function prints a message indicating that the filename is blank.
}
\examples{
\dontrun{
print(get_current_file_location())
}
}
