% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martys_table_style.r,
%   R/martys_table_style_legacy_huxtable.r
\name{martys_table_style_legacy_huxtable}
\alias{martys_table_style_legacy_huxtable}
\title{Apply style to a given table, return a huxtable (or kableExtra table)}
\usage{
martys_table_style_legacy_huxtable(
  tabl,
  alig. = c("c"),
  caption. = "",
  scale_down. = FALSE,
  fontsize. = 7,
  landscape. = TRUE
)

martys_table_style_legacy_huxtable(
  tabl,
  alig. = c("c"),
  caption. = "",
  scale_down. = FALSE,
  fontsize. = 7,
  landscape. = TRUE
)
}
\arguments{
\item{tabl}{The table to which the style should be applied.}

\item{alig.}{The alignment to be applied to the table. Default is "c" for center.}

\item{caption.}{The caption of the table. Default is the result of generate_tab_title().}

\item{scale_down.}{Logical, if TRUE, scales down the table if it is wider than the text width. Default is FALSE.}

\item{fontsize.}{The font size to be applied to the table. Default is 7.}

\item{landscape.}{Logical, if TRUE, the table will be displayed in landscape format. Default is TRUE.}
}
\value{
A table with the specified style applied.

A table with the specified style applied.
}
\description{
This function applies a particular style to a table. The table can be a data frame,
a gtsummary table, or a knitr kable. Different processing steps are applied based
on the class of the input table.

This function applies a particular style to a table. The table can be a data frame,
a gtsummary table, or a knitr kable. Different processing steps are applied based
on the class of the input table.
}
\examples{
\dontrun{
data(mtcars)
martys_table_style(mtcars)
}
\dontrun{
data(mtcars)
martys_table_style_legacy_huxtable(mtcars)
}
}
