% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_recent_files.r
\name{move_recent_files}
\alias{move_recent_files}
\title{Move Recent Files from One Directory to Another}
\usage{
move_recent_files(project_dir, folder, delay_hs = 1)
}
\arguments{
\item{project_dir}{A character string specifying the root directory of the project.}

\item{folder}{A character string specifying the name of the source directory (without path), from which files are to be moved.
For example: "\if{html}{\out{<folder>}}_files/figure-latex".}

\item{delay_hs}{Amount of time in which the files are considered 'recent' in hours. Defaults to '1'.}
}
\value{
This function does not return a value. It is called for its side effect of moving files.
If no files are found in the source directory, the function will stop and return an error message.
Unfortunately it deletes only the last folder.
}
\description{
This function moves files from one directory to another that were modified within the last hour.
It is specifically designed to work within a project directory structure, taking files (figures) from a source
directory and moving them to a "vignettes" directory within the project.
}
\note{
This function requires the 'lubridate' package to determine which files are less than an hour old.
}
\examples{
\dontrun{
move_recent_files("path/to/project", "myfolder")
}

}
