% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompile_quick_vignette.r
\name{precompile_quick_vignette}
\alias{precompile_quick_vignette}
\title{Precompiles a quick vignette from a computationally expensive report}
\usage{
precompile_quick_vignette(input_file)
}
\arguments{
\item{input_file}{The name of the .Rmd file to be precompiled (must include the ".rmd" extension)}
}
\value{
Invisible NULL; the output will be in the same directory as the source named ..._quick.Rmd
}
\description{
This function takes an R Markdown file that generates a computationally expensive report
and turns it into a quicker markdown document.
}
\examples{
\dontrun{
precompile_quick_vignette("example_report.rmd")
}
}
\references{
https://ropensci.org/blog/2019/12/08/precompute-vignettes/
}
