% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\docType{data}
\name{theme_default_ggplot}
\alias{theme_default_ggplot}
\title{Default Theme for ggplot}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 97.
}
\usage{
theme_default_ggplot
}
\value{
A ggplot theme object that can be added to a ggplot.
}
\description{
This function applies a pre-specified theme to a ggplot object. The theme
is based on the 'theme_bw' function from the 'ggplot2' package and adds
some modifications. Specifically, it removes the legend, adjusts the plot
caption, and changes the font family to "serif".
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(mpg, cyl)) + geom_point()
p + theme_default_ggplot()
}
}
\keyword{datasets}
