% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompile_quick_vignette.r
\name{precompile_quick_vignette}
\alias{precompile_quick_vignette}
\title{Precompiles a quick vignette from a computationally expensive report}
\usage{
precompile_quick_vignette(file_name)
}
\arguments{
\item{file_name}{The name of the .Rmd file to be precompiled (must include the ".rmd" extension)}
}
\value{
Invisible NULL
}
\description{
This function takes an R Markdown file that generates a computationally expensive report
and turns it into a quicker markdown document. This markdown document is intended to be
turned into a vignette. This is achieved by copying the .Rmd file, renaming it to .Rmd.orig,
and knitting it with the output sent to a "vignettes/" directory. The function will also
move any figures generated in the project directory to the "vignettes/" directory.
If a PDF of the report is available, it will also be copied to the "docs/articles/" directory.
}
\examples{
\dontrun{
precompile_quick_vignette("example_report.rmd")
}
}
\references{
https://ropensci.org/blog/2019/12/08/precompute-vignettes/
}
